backend/requirements.txt
# Web
fastapi==0.111.0
uvicorn==0.30.0

# Data
sqlalchemy==2.0.30
psycopg2-binary==2.9.9
pgvector==0.3.0
alembic==1.13.0

# PDF
pdfplumber==0.11.0
pymupdf==1.24.0

# AWS
boto3==1.34.0

# LLM + Embeddings (OpenAI)
openai==1.30.0

# Auth
python-jose[cryptography]==3.3.0
bcrypt==4.1.3
python-multipart==0.0.9

pydantic-settings==2.2.0
python-dotenv==1.0.0

httpx==0.27.2
